Removed share page
