+++ app/src/main/java/com/example/snapchat/MusicFragment.java
@@ -1,7 +1,9 @@
 package com.example.snapchat;
+import android.Manifest;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.hardware.Camera;
+import android.media.MediaPlayer;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
@@ -19,12 +21,15 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageView;
+import android.widget.ListAdapter;
+import android.widget.ListView;
 import android.widget.PopupMenu;
 import android.widget.Toast;
 
 import com.google.firebase.auth.FirebaseAuth;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 public class MusicFragment extends Fragment {
@@ -34,18 +39,44 @@ public class MusicFragment extends Fragment {
         return fragment;
     }
 
+   // ListView listView;
+    List<String> list;
+    String str;
+   // ListAdapter listAdapter;
+    MediaPlayer mediaPlayer;
+
+    Button back;
     @Nullable
     @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+    public View onCreateView( LayoutInflater inflater,  ViewGroup container,  Bundle savedInstanceState) {
 
         View view = inflater.inflate(R.layout.fragment_music,container,false);
-        EditText v = view.findViewById(R.id.EnterSong);
-        v.setText(Html.fromHtml(v.getText().toString()));
-        v.setMovementMethod(LinkMovementMethod.getInstance());
+        //Bundle p = getActivity().getIntent().getExtras();
+        //ShowMusicActivity activity=(ShowMusicActivity)getActivity();
+      //  Bundle bundle=this.getArguments();
+   //   final ArrayList<String> arrayList;
+      //if(bundle!=null)
+    //      arrayList=bundle.getStringArrayList("info");
+//       // if (getArguments() != null) {
+ //          arrayList = this.getArguments().getStringArrayList("info");
+//      //  }
+//         System.out.println(arrayList);
+//        //final ArrayList<String> arrayList =p.getStringArrayList("info");
+//        ///System.out.println(arrayList);
+      //  EditText v = view.findViewById(R.id.song1);
+     //   v.setText(Html.fromHtml(v.getText().toString()));
+    //    v.setMovementMethod(LinkMovementMethod.getInstance());
         ImageView mPost = view.findViewById(R.id.post);
         Button mlogout = view.findViewById(R.id.logout);
         //Button mPost = view.findViewById(R.id.post);
 //        Button mCapture= view.findViewById(R.id.capture);
+
+
+     //   EditText artist=view.findViewById(R.id.artist1);
+
+      //  v.setText(arrayList.get(0));
+     //   artist.setText(arrayList.get(1));
+        CheckPermission();
         mlogout.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
@@ -69,6 +100,15 @@ public class MusicFragment extends Fragment {
 
     private void FindUsers() {
 
+    }
+
+    private void CheckPermission(){
+
+        if (ActivityCompat.checkSelfPermission(getContext(), Manifest.permission.READ_EXTERNAL_STORAGE) !=
+                PackageManager.PERMISSION_GRANTED) {
+            requestPermissions(new String[]{Manifest.permission.READ_EXTERNAL_STORAGE}, 123);
+            return;
+        }
     }
 
     private void MusicPlayer(){
