--- app/src/main/res/layout/activity_share.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".Share">
-    <ListView
-        android:id="@+id/listView"
-        android:layout_width="match_parent"
-        android:layout_height="90dp"></ListView>
-    <LinearLayout
-
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="300dp"
-        android:layout_centerHorizontal="true"
-        android:orientation="vertical">
-    <EditText
-        android:id="@+id/song"
-        android:hint="Enter Song Name"
-        android:layout_centerHorizontal="true"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-
-        />
-    <EditText
-        android:id="@+id/artist"
-        android:hint="Enter Artist Name"
-        android:layout_marginTop="50dp"
-        android:layout_width="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_height="wrap_content"
-
-        />
-    </LinearLayout>
-    <Button
-        android:id="@+id/back"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-
-        android:text="Go Back!"
-        android:layout_centerHorizontal="true"
-        android:layout_alignParentBottom="true"/>
-
-</RelativeLayout>
