+++ app/src/main/res/layout/fragment_music.xml
@@ -34,7 +34,7 @@
         android:layout_centerHorizontal="true"
         android:orientation="vertical">
         <EditText
-            android:id="@+id/song"
+            android:id="@+id/song1"
             android:hint="Enter Song Name"
 
             android:layout_width="wrap_content"
@@ -42,7 +42,7 @@
             android:textColorHint="@color/yellow"
             />
         <EditText
-            android:id="@+id/artist"
+            android:id="@+id/artist1"
             android:hint="Enter Artist Name"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
